[booster-browser] Add browser specific booster. Fixes JB#24329
This a fork from silica booster. The browser booster
depends on QtMozEmbed so that we get the jemalloc implementation
in place.

QML preloading is minimized and optimized for browser use. None
of the preloaded components is initialized, only compiled. This
reduces memory usage of the browser booster by 2megs.

With Silica preload.qml:
Resident 22MB and shared 13MB => 9MB

With custom preload.qml:
Resident 17MB and shared 10MB => 7MB